+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   370](https://github.com/mrkkrp/megaparsec/issues/370).
 * Inlined `Applicative` operators `(<*)` and `(*>)`. [PR
   566](https://github.com/mrkkrp/megaparsec/pull/566).
+* `many` and `some` of the `Alternative` instance of `ParsecT` are now more
+  efficient, since they use the monadic implementations under the hood.
+  [Issue 567](https://github.com/mrkkrp/megaparsec/issues/567).
 
 ## Megaparsec 9.6.1